+++ src/AiRecommendation/pages/AiRecommendationLoading.tsx
@@ -1,0 +1,13 @@
+import React, { useState, useEffect } from 'react';
+import Layout from '../layout/Layout.tsx';
+
+
+const AiRecommendationLoading = () => {
+  return (
+    <>
+      <Layout />
+    </>
+  );
+};
+
+export default AiRecommendationLoading;
