+++ src/App.tsx
@@ -1,21 +1,9 @@
-import {BrowserRouter, Routes, Route} from "react-router-dom"
+import Router from "./route/Router.tsx"
 import './App.css';
-import SurveyController from "./pages/SurveyController.tsx";
-import Home from "./pages/AiRecommendationLoading.tsx"; // 임시
-import Result from "./pages/AiRecommendationResult.tsx" // 임시
-import SurveyComplete from "./pages/SurveyComplete.tsx";
 
 const App = () => {
   return (
-    <div className="App">
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<SurveyController />}/>
-          <Route path="/survey" element={<SurveyController />}/>
-          <Route path="/result" element={<Result />}/> 
-        </Routes>
-      </BrowserRouter>
-    </div>
+    <Router />
   );
 }
