+++ src/Survey/components/ProgressBar/ProgressBar.css
@@ -1,0 +1,19 @@
+/* Progress Bar */
+.survey-progress-bar {
+  padding: 0.5rem 1.25rem 1.75rem 1.25rem; /* 8px 20px 28px 20px*/
+  z-index: -100;
+}
+
+.survey-progress-bar-background {
+  width: 100%;
+  background-color: #0000000F;
+  border-radius: 9999px; /* 완전한 캡슐형 */
+  height: 0.375rem; /* 6px */
+}
+
+.survey-progress {
+  background: linear-gradient(90deg, #CBFF69 0%, #2DC5F4 100%);
+  height: 0.375rem; /* 6px */
+  border-radius: 9999px; /* 완전한 캡슐형 */
+  transition: all 0.5s cubic-bezier(0.4, 0, 0.2, 1); /* transition-all duration-500 ease-out */
+}
