+++ src/Survey/components/ProgressBar/ProgressBar.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import "./ProgressBar.css"
+
+const ProgressBar: React.FC<ProgressBarProps> = ({ progress }) => {
+    return (
+    <div className="survey-progress-bar">
+        <div className="survey-progress-bar-background">
+        <div 
+            className="survey-progress"
+            style={{ width: `${progress}%` }}
+        ></div>
+        </div>
+    </div> 
+    );
+}
+
+export default ProgressBar;
