+++ src/Survey/layout/Header.css
@@ -1,0 +1,27 @@
+/* Header */
+.survey-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 1rem;
+  background-color: white;
+  z-index: -100;
+}
+
+.survey-arrow-left {
+  cursor: pointer;
+  width: 1.75rem; /*28px*/
+  height: 1.75rem;
+}
+
+.survey-title {
+  font-size: 1.125rem; /*18px*/
+  font-weight: 600;
+  line-height: 130%;
+  letter-spacing: -2.5%;
+}
+
+.survey-header-temp {
+    width: 1.75rem; /*28px*/
+    height: 1.75rem;
+}
