+++ src/Survey/layout/style.css
@@ -8,52 +8,6 @@
   min-height: 100vh;
 }
 
-/* Header */
-.survey-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 1rem;
-  background-color: white;
-}
-
-.survey-arrow-left {
-  cursor: pointer;
-  width: 1.75rem; /*28px*/
-  height: 1.75rem;
-}
-
-.survey-title {
-  font-size: 1.125rem; /*18px*/
-  font-weight: 600;
-  line-height: 130%;
-  letter-spacing: -2.5%;
-}
-
-.survey-header-temp {
-    width: 1.75rem; /*28px*/
-    height: 1.75rem;
-}
-
-/* Progress Bar */
-.survey-progress-bar {
-  padding: 0.5rem 1.25rem 1.75rem 1.25rem; /* 8px 20px 28px 20px*/
-}
-
-.survey-progress-bar-background {
-  width: 100%;
-  background-color: #0000000F;
-  border-radius: 9999px; /* 완전한 캡슐형 */
-  height: 0.375rem; /* 6px */
-}
-
-.survey-progress {
-  background: linear-gradient(90deg, #CBFF69 0%, #2DC5F4 100%);
-  height: 0.375rem; /* 6px */
-  border-radius: 9999px; /* 완전한 캡슐형 */
-  transition: all 0.5s cubic-bezier(0.4, 0, 0.2, 1); /* transition-all duration-500 ease-out */
-}
-
 /* Content */
 .survey-contents {
   flex: 1;
