+++ src/Survey/pages/Survey.tsx
@@ -1,0 +1,11 @@
+import Layout from "../layout/Layout.tsx";
+
+const Survey: React.FC<SurveyProps> = ({ onComplete }) => {
+    return (
+        <>
+            <Layout onComplete={onComplete}/>
+        </>
+    )
+}
+
+export default Survey;
