+++ src/SurveyComplete/layout/style.css
@@ -1,0 +1,56 @@
+/* SurveyComplete.css */
+.survey-complete-container {
+  margin: 0 auto;
+  background-color: #00000040;
+
+  max-width: 375px;
+  min-height: 100vh;
+
+  display: flex;
+  flex-direction: column;
+
+  z-index: 100;
+}
+
+.survey-complete-content {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    padding: 2rem;
+
+    margin-top: -100px;
+}
+
+.survey-complete-content-title {
+    font-size: 4.5rem; /* 72px */
+    font-weight: 700;
+    margin-bottom: 2rem; /* 32px */
+    animation: fadeInUp 0.6s ease-out;
+}
+
+.survey-complete-content-subtitle {
+    font-size: 2.5rem; /* 40px */
+    font-weight: 600;
+    margin-top: 0;
+    animation: fadeInUp 0.6s ease-out 0.2s both;
+}
+
+@keyframes fadeInUp {
+  0% {
+    opacity: 0;
+    transform: translateY(20px);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+@media (max-width: 768px) {
+    .survey-complete-container {
+        max-width: 100%;
+  }
+}
