+++ src/SurveyComplete/pages/SurveyComplete.tsx
@@ -1,0 +1,12 @@
+// components/SurveyComplete.tsx
+import Layout from "../layout/Layout.tsx"
+
+const SurveyComplete = () => {
+  return (
+    <>
+      <Layout />
+    </>
+  );
+};
+
+export default SurveyComplete;
