--- src/components/Dealer_Carousel.css
@@ -1,130 +1,0 @@
-.dealer-carousel {
-    display: flex;
-    width: 120px;
-    flex-direction: column;
-    align-items: flex-start;
-    flex-shrink: 0;
-    position: relative;
-}
-
-.dealer-carousel-img {
-    display: flex;
-    height: 152px;
-    padding: 0px 4px;
-    justify-content: center;
-    align-items: flex-end;
-    gap: 10px;
-    align-self: stretch;
-    border-radius: 12px 12px 0px 0px;
-    background: #fff;
-    
-    box-shadow: 0px 0px 4px 0px rgba(0, 0, 0, 0.16);
-}
-
-.dealer-carousel-img img {
-    width: 112px;
-    height: 144px;
-    border-radius: 12px 12px 0px 0px;
-    background: lightgray -4px -4.898px / 107.143% 122.449% no-repeat;
-
-    object-fit: cover;
-}
-
-.dealer-carousel-rating {
-    display: flex;
-    align-items: center;
-    gap: 4px;
-
-    position: absolute;
-    right: 8px;
-    top: 4px;
-}
-
-.dealer-carousel-rating img{
-    width: 16px;
-    height: 16px;
-    aspect-ratio: 1/1;
-}
-
-.dealer-carousel-rating p{
-    width: 24px;
-    height: 24px;
-    color: #2DC5F4;
-    font-family: Pretendard, sans-serif;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 150%; /* 24px */
-    letter-spacing: -0.352px;
-
-    margin: 0px;
-}
-
-.dealer-carousel-text {
-    display: flex;
-    flex-direction: column;
-    width: 120px;
-    height: 48px;
-    flex-shrink: 0;
-    border-radius: 0px 0px 12px 12px;
-    background: #AEFF1C;
-    padding: 0px 8px;
-}
-
-.dealer-name-line {
-    display: flex;
-    flex-direction: row;
-    align-items: flex-end;
-}
-
-.dealer-name-line p:nth-child(1) {
-    width: 56px;
-    color: #FFF;
-    text-shadow: 1px 1px 4px rgba(0, 0, 0, 0.25);
-    font-family: Pretendard, sans-serif;
-    font-size: 20px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 150%; /* 30px */
-    letter-spacing: -0.44px;
-
-    margin: 0px;
-
-    text-align: start;
-}
-
-.dealer-name-line p:nth-child(2) {
-    display: flex;
-    width: 24px;
-    height: 16px;
-    flex-direction: column;
-    justify-content: flex-end;
-    flex-shrink: 0;
-
-    color: #FFF;
-    text-shadow: 1px 1px 4px rgba(0, 0, 0, 0.25);
-    font-family: Pretendard, sans-serif;
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 150%; /* 18px */
-    letter-spacing: -0.264px;
-
-    margin: 0px;
-}
-
-.dealer-carousel-text p {
-    width: 104px;
-    color: #FFF;
-    text-align: right;
-    text-shadow: 1px 1px 4px rgba(0, 0, 0, 0.25);
-    font-family: Pretendard;
-    font-size: 10px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 150%; /* 15px */
-    letter-spacing: -0.22px;
-
-    margin: 0px;
-}
-
