--- src/pages/SurveyComplete.css
@@ -1,108 +1,0 @@
-/* SurveyComplete.css */
-.survey-complete-container {
-  margin: 0 auto;
-  background-color: #00000040;
-
-  max-width: 375px;
-  min-height: 100vh;
-
-  display: flex;
-  flex-direction: column;
-}
-
-/* Header */
-.survey-complete-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 1rem;
-}
-
-.survey-complete-arrow-left {
-  cursor: pointer;
-  width: 1.75rem; /*28px*/
-  height: 1.75rem;
-}
-
-.survey-complete-title {
-  font-size: 1.125rem; /*18px*/
-  font-weight: 600;
-  line-height: 130%;
-  letter-spacing: -2.5%;
-}
-
-.survey-complete-header-temp {
-    width: 1.75rem; /*28px*/
-    height: 1.75rem;
-}
-
-/* Progress Bar */
-.survey-complete-progress-bar {
-  padding: 0.5rem 1.25rem 1.75rem 1.25rem; /* 8px 20px 28px 20px*/
-}
-
-.survey-complete-progress-bar-background {
-  width: 100%;
-  background-color: #0000000F;
-  border-radius: 9999px; /* 완전한 캡슐형 */
-  height: 0.375rem; /* 6px */
-}
-
-.survey-complete-progress {
-    background: linear-gradient(90deg, 
-    rgba(203, 255, 105, 0.3) 0%, 
-    rgba(45, 197, 244, 0.3) 100%); /* 투명도로 흐리게 */
-    height: 0.375rem; /* 6px */
-    border-radius: 9999px;
-    width: 100%; /* 완료 상태이므로 100% */
-    transition: all 0.5s cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-.survey-complete-content {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-    padding: 2rem;
-
-    margin-top: -100px;
-}
-
-.survey-complete-content-title {
-    font-size: 4.5rem; /* 72px */
-    font-weight: 700;
-    margin-bottom: 2rem; /* 32px */
-    animation: fadeInUp 0.6s ease-out;
-}
-
-.survey-complete-content-subtitle {
-    font-size: 2.5rem; /* 40px */
-    font-weight: 600;
-    margin-top: 0;
-    animation: fadeInUp 0.6s ease-out 0.2s both;
-}
-
-/* 애니메이션 */
-@keyframes progressComplete {
-  0% { width: 0%; }
-  100% { width: 100%; }
-}
-
-@keyframes fadeInUp {
-  0% {
-    opacity: 0;
-    transform: translateY(20px);
-  }
-  100% {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-@media (max-width: 768px) {
-    .survey-complete-container {
-        max-width: 100%;
-  }
-}
